move DrawerContent component to it's own folder
